Moved scrollIntoView.js into utils folder
